update listener interface
