+++ ava/components.py
@@ -23,11 +23,12 @@ class ComponentManager(object):
         self._components = []
         self._config_keys = {}
         self._queues = {}
+        self._queues['QueueWindowsListener'] = Queue()
         self._queues['QueueComponentManager'] = Queue()
         self._config = ConfigLoader(path.dirname(path.realpath(__file__)), self._queues)
         self._config.load('settings.json')
-        if platform.system() == 'Windows':
-            self._queues['QueueWindowsListener'] = Queue()
+        # if platform.system() == 'Windows':
+        #     self._queues['QueueWindowsListener'] = Queue()
 
 
     def add_component(self, Component):
