+++ ava/plugins/listener/platforms/unix.py
@@ -1,48 +1,14 @@
-import ast
 import select
-from ...process import flush_stdout
 from .interface import _ListenerInterface
-from ...process import multi_lines_output_handler
-from avasdk.plugins.log import Logger
 
 class _UnixInterface(_ListenerInterface):
     """
     """
 
-    def __init__(self, state, store, tts):
+    def __init__(self, state, store, tts, listener):
         """
         """
-        super().__init__(state, store, tts)
-
-    def _process_result(self, plugin_name, process):
-        """This functions flushes the stdout of the given process and process the
-            data read.
-
-        params:
-            - plugin_name: The name of the plugin (string).
-            - process: The process object (subprocess.Popen)
-        """
-        output, import_flushed = flush_stdout(process)
-        if Logger.ERROR in output:
-            self.queue_tts.put('Plugin {} just crashed... Restarting'.format(plugin_name))
-            self.store.get_plugin(plugin_name).kill()
-            self.store.get_plugin(plugin_name).restart()
-            return
-        if Logger.IMPORT in output:
-            return
-        if Logger.REQUEST in output:
-            output.remove(Logger.REQUEST)
-            self.state.plugin_requires_user_interaction(plugin_name)
-            self.queue_tts.put(ast.literal_eval(''.join(output)).get('tts'))
-            return
-        output.remove(Logger.RESPONSE)
-        result, multi_lines = multi_lines_output_handler(output)
-        if multi_lines:
-            # TODO Spawn a window and print the multi lines output
-            print(result)
-            self.queue_tts.put('Result of [{}] has been print.'.format(plugin_name))
-        else:
-            self.queue_tts.put(result)
+        super().__init__(state, store, tts, listener)
 
     def listen(self):
         """
@@ -60,9 +26,3 @@ class _UnixInterface(_ListenerInterface):
                 continue
             p = OBSERVED.get(fd)
             self._process_result(''.join('{}'.format(k) for k, v in self.store.plugins.items() if p == v.get_process()), p)
-
-    def stop(self):
-        """
-        """
-        for _, plugin in self.store.plugins.items():
-            plugin.get_process().stdout.close()
