+++ ava/plugins/process/python_main.py
@@ -51,7 +51,6 @@ def execute(plugin_name, command):
         - plugin_name: The name of the plugin running (string).
         - command: the user's input (string).
     """
-    # print('python_main: ', command)
     if PLUGIN.get(plugin_name):
         command_name, args = split_string(command, ' ')
         plugin = PLUGIN.get(plugin_name)
@@ -66,9 +65,8 @@ def execute(plugin_name, command):
 if __name__ == "__main__":
     try:
         plugin_name = sys.argv[1]
-        # print(sys.argv)
         import_module(plugin_name)
         wait_for_command(plugin_name)
-    except Exception as err:
-        print(str(err))
+    except:
+        # TODO log to a file
         Logger.log_error()
